pyproject.toml
[project]
name = "nltbuild"
version = "1.6.39"
description = "nltbuild"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [ "uv>=0.8.4", "ruff>=0.12.7", "toml>=0.10.2", "typer-slim", "nltlog>=1.0.11", "funshell>=1.0.2",]
keywords = [ "build", "requirements", "packaging", "uv",]
[[project.authors]]
name = "牛哥"
email = "niuliangtao@qq.com"

[[project.authors]]
name = "farfarfun"
email = "farfarfun@qq.com"

[[project.maintainers]]
name = "牛哥"
email = "niuliangtao@qq.com"

[[project.maintainers]]
name = "farfarfun"
email = "farfarfun@qq.com"

[project.license]
text = "MIT"

[project.scripts]
nltbuild = "nltbuild.core:nltbuild"

[project.urls]
Organization = "https://github.com/farfarfun"
Repository = "https://github.com/farfarfun/funbuild"
Releases = "https://github.com/farfarfun/funbuild/releases"

[tool.ruff]
indent-width = 4
line-length = 120

[tool.setuptools]
license-files = []

[tool.uv]
package = true

[tool.ruff.lint]
select = [ "I", "E", "W", "B", "F", "UP", "RUF", "PL",]
fixable = [ "I",]
ignore = [ "E501", "UP006", "UP007", "PLR0911", "PLR0912", "PLR0913", "PLR0915", "PLR2004",]
